Prevent multiple triggers of cascading callbacks on children
This was a struggle to fit into the existing codebase.. The
method to get cascading children returns all of them, not just
the ones for the current level, and is called by each level of the
hierarchy in run_callbacks. This works around it (in an ugly way)
so that cascading children are fetched and processed only once in
the initial call to run_callbacks and the super method is used for
each child + parent.